* Need cast to double.

+++ src/nlua_pilot.c
@@ -1058,7 +1058,7 @@ static int pilotL_weapset( lua_State *L )
          if ((outfit_isLauncher(slot->outfit) ||
                   outfit_isFighterBay(slot->outfit)) &&
                (slot->u.ammo.outfit != NULL)) {
-            lua_pushnumber( L, slot->u.ammo.quantity / outfit_amount(slot->outfit) );
+            lua_pushnumber( L, (double)slot->u.ammo.quantity / (double)outfit_amount(slot->outfit) );
          }
          else
             lua_pushnil( L );
